requirements.txt
# HCR - Hybrid Cognitive Runtime
# Core
requests>=2.28.0
python-dotenv>=1.0.1
watchdog>=4.0.0

# LLM Providers (install the one(s) you need)
groq>=0.11.0              # Default: free tier, fast
google-genai>=1.0.0       # Fallback: free tier
# ollama has no SDK dependency (uses urllib)
